\name{corrHeatMap}
\alias{corrHeatMap}
\title{Correlation Heat Map}
\usage{
corrHeatMap(df,roundto=2,na.rm=FALSE,plotText=TRUE)
}
\description{
Creates correlation matrix from data and:
1) prints top ten correlations as output and as plot in form of heat map
2) returns all correlations as a tibble.
}

