\name{rankedCorrelations}
\alias{rankedCorrelations}
\title{Return matched substrings from preset character vector of terms}
\usage{
rankedCorrelations(data)
}
\description{
Returns data frame with variable pairs ordered by absolute correlation. Takes continuous variables and  categorical variables in either charactor or factor classes.
