\name{corrBars}
\alias{CorrBars}
\title{Correlation Bar Plot}
\usage{
corrBars(df,roundto=2,na.rm=FALSE)
}
\description{
Creates correlation matrix from data and:
1) prints top ten correlations as output and as plot in form of a bar plot
2) returns correlation matrix; all ordered correlation pairs as a tibble.
}

